Change date format for DatePickerInput.
This is a work around.  I was seeing a date validation error when I entered
the date 24/7/2023. In US date format that's not a valid date. But my
browser has the date format in DD/MM/YYYY, and the date format in this
DatePickerInput is also that format. So 24/7/2023 should have been a valid
date. Changing to yyyy-mm-dd made the validation error stop occurring,
though I don't really understand why.